Split oversized files to satisfy CI line-count gate
prompt_execution.rs (2604 lines) is split by moving the 1117-line tool
call execution block into prompt_execution_parts/tool_processing.rs,
included via the include! macro as a block expression. The main file
drops to 1488 lines.

suite_a.rs (2489 lines) is split by moving test helpers to tests.rs
(accessible to all suites via use super::*) and extracting the
context-budget tests plus normalize_tool_args tests into suite_c.rs.
The main suite drops to 1705 lines; suite_c.rs is 555 lines.

All engine_loop tests pass. CI file-size-check reports all touched
files within limits.
